[feat] 캘린더 통신 종료

+++ app/src/main/java/com/example/daon/conect/calendar/BodyListInsertRequestDto.kt
@@ -1,10 +1,11 @@
 package com.example.daon.conect.calendar
 
+import android.widget.EditText
 import com.google.gson.annotations.SerializedName
 
 data class BodyListInsertRequestDto(
-    @SerializedName("temperature") val temperature : Double,
-    @SerializedName("weight") val weight : Int,
-    @SerializedName("fastingBloodSugar") val fastingBloodSugar : Int,
-    @SerializedName("note") val note : String
+    @SerializedName("temperature") val temperature: Double,
+    @SerializedName("weight") val weight: Double,
+    @SerializedName("fastingBloodSugar") val fastingBloodSugar: Int,
+    @SerializedName("note") val note: String
 )
